[GH-#335] General code refactoring, RLP encodings now are in corresponding structure modules
