+++ apps/aecore/lib/aecore/account/account.ex
@@ -13,7 +13,7 @@ defmodule Aecore.Account.Account do
   alias Aecore.Tx.DataTx
   alias Aecore.Tx.SignedTx
   alias Aecore.Account.AccountStateTree
-
+  alias Aeutil.Serialization
   @type t :: %Account{
           balance: non_neg_integer(),
           nonce: non_neg_integer(),
@@ -149,4 +149,42 @@ defmodule Aecore.Account.Account do
   def base58c_decode(bin) do
     {:error, "#{__MODULE__}: Wrong data: #{inspect(bin)}"}
   end
+  @spec rlp_encode(Account.t(), Wallet.pubkey()) :: binary()
+  def rlp_encode(%Account{} = account, pkey) when is_binary(pkey) do
+    [
+      type_to_tag(Account),
+      get_version(Account),
+      pkey, #pubkey ,
+      account.nonce, #nonce
+      account.last_updated,    #height
+      account.balance #balance
+    ]
+    |>
+    ExRLP.encode
+  end
+  def rlp_encode(_) do
+    {:error, :illegal_serialization}
+  end
+  def rlp_decode(values) when is_binary(values) do
+    [tag_bin, ver_bin | rest_data] = ExRLP.decode(values)
+    tag = Serialization.transform_item(tag_bin, :int)
+    ver = Serialization.transform_item(ver_bin, :int)
+    case tag_to_type(tag) do
+      Account ->
+    [pkey, nonce, height, balance] = rest_data
+    [pkey, Serialization.transform_item(nonce, :int), Serialization.transform_item(height, :int), Serialization.transform_item(balance, :int)]
+     {:ok, %Account{balance: Serialization.transform_item(balance, :int), last_updated: Serialization.transform_item(height, :int), nonce: Serialization.transform_item(nonce, :int)}}
+      _ -> {:error, :invalid_serialization}
+    end
+  end
+  def rlp_decode(:none) do
+    :none
+  end
+  def rlp_decode(_) do
+    {:error, :illegal_serialization}
+  end
+
+  defp type_to_tag(Account), do: 10
+  defp tag_to_type(10), do: Account
+  defp get_version(Account), do: 1
 end
