+++ apps/aecore/lib/aecore/tx/signed_tx.ex
@@ -38,7 +38,7 @@ defmodule Aecore.Tx.SignedTx do
   """
   @spec validate(SignedTx.t()) :: :ok | {:error, String.t()}
   def validate(%SignedTx{data: data} = tx) do
-    if Signing.verify(Serialization.rlp_encode(data), tx.signature, data.sender) do
+    if Signing.verify(DataTx.rlp_encode(data), tx.signature, data.sender) do
       :ok
     else
       {:error, "#{__MODULE__}: Can't verify the signature
@@ -58,7 +58,7 @@ defmodule Aecore.Tx.SignedTx do
   """
   @spec sign_tx(DataTx.t(), binary()) :: {:ok, SignedTx.t()}
   def sign_tx(%DataTx{} = tx, priv_key) when byte_size(priv_key) == 32 do
-    signature = Signing.sign(Serialization.rlp_encode(tx), priv_key)
+    signature = Signing.sign(DataTx.rlp_encode(tx), priv_key)
 
     if byte_size(signature) <= get_sign_max_size() do
       {:ok, %SignedTx{data: tx, signature: signature}}
@@ -81,7 +81,7 @@ defmodule Aecore.Tx.SignedTx do
 
   @spec hash_tx(SignedTx.t()) :: binary()
   def hash_tx(%SignedTx{data: data}) do
-    Hash.hash(Serialization.rlp_encode(data))
+    Hash.hash(DataTx.rlp_encode(data))
   end
 
   @spec reward(DataTx.t(), Account.t()) :: Account.t()
@@ -128,4 +128,52 @@ defmodule Aecore.Tx.SignedTx do
   def base58c_decode_signature(_) do
     {:error, "Wrong data"}
   end
+  
+  @spec rlp_encode(DataTx.t(SignedTx.t())) :: binary() | atom()
+  def rlp_encode(%SignedTx{} = tx) do
+      signatures = for sig <- [tx.signature] do
+        if sig == nil do  #workaround - should be removed when CoinbaseTx will have its own structure
+          ExRLP.encode(<<0>>)
+        else
+          ExRLP.encode(sig)
+        end
+      end
+    ExRLP.encode([type_to_tag(SignedTx), get_version(SignedTx), signatures, DataTx.rlp_encode(tx.data)])
+  end
+  def rlp_encode(_) do
+    :invalid_signedtx
+  end
+
+  @spec rlp_decode(binary()) :: SignedTx.t() | atom()
+  def rlp_decode(values) when is_binary(values) do
+    [tag_bin, ver_bin | rest_data] = ExRLP.decode(values)
+    tag = Serialization.transform_item(tag_bin, :int)
+    ver = Serialization.transform_item(ver_bin, :int)
+    case tag_to_type(tag) do
+      SignedTx -> 
+    [signatures, tx_data] = rest_data
+    decoded_signatures = 
+      for sig <- signatures do
+        ExRLP.decode(sig)
+      end
+    %SignedTx{data: rlp_decode(tx_data) , signature: decoded_signatures}
+    _ -> :invalid_serialization
+    end
+  end
+  def rlp_decode(_) do
+    :invalid_serialization
+  end
+   
+  @spec type_to_tag(atom()) :: integer() | atom()
+  defp type_to_tag(SignedTx), do: 11
+  defp type_to_tag(_), do: :unknown_type
+  
+  @spec tag_to_type(integer()) :: SignedTx | atom()
+  defp tag_to_type(11), do: SignedTx
+  defp tag_to_type(_), do: :unknown_tag
+  
+  @spec get_version(SignedTx) :: integer() | atom()
+  defp get_version(SignedTx), do: 1
+  defp get_version(_), do: :unknown_struct_version
+
 end
