+++ apps/aeutil/lib/serialization.ex
@@ -315,468 +315,12 @@ defmodule Aeutil.Serialization do
   defp serialize_txs_info_to_json([], acc) do
     Enum.reverse(acc)
   end
-@spec rlp_encode(DataTx.t(SignedTx.t())) :: binary()
-  def rlp_encode(%SignedTx{} = tx) do
-      signatures = for sig <- [tx.signature] do
-        if sig == nil do  #workaround - should be removed when CoinbaseTx will have its own structure
-          ExRLP.encode(<<0>>)
-        else
-          ExRLP.encode(sig)
-        end
-      end
-    ExRLP.encode([type_to_tag(SignedTx), 1, signatures, rlp_encode(tx.data)])
-  end
-  @spec rlp_encode(DataTx.t(SpendTx.t())) :: binary()
-  def rlp_encode(%DataTx{type: SpendTx} = tx) do
 
-    if tx.sender == nil do
-      [
-        type_to_tag(CoinbaseTx),
-        get_version(CoinbaseTx),
-        tx.payload.receiver, #receiver
-        # Subject to discuss/change:
-        # CoinbaseTx Should have a "height" field, currently "nonce" is being encoded
-        tx.nonce, #nonce / but should be height
-        tx.payload.amount #reward
-      ]
-    else
-      [
-        type_to_tag(SpendTx),
-        get_version(SpendTx),
-        tx.sender, # sender
-        tx.payload.receiver, #receiver
-        tx.payload.amount, #amount
-        tx.fee, # fee
-        tx.nonce # nonce
-      ]
-    end
-    |> ExRLP.encode()
-  end
-  @spec rlp_encode(DataTx.t()) :: binary()
-  def rlp_encode(%DataTx{type: CoinbaseTx} = tx) do
-
-    [
-      type_to_tag(CoinbaseTx),
-      get_version(CoinbaseTx),
-      tx.payload.receiver, #receiver
-      # Subject to discuss/change:: Here should be Height, but at this moment nonce is being encoded
-      tx.nonce, # nonce / but should be height
-      tx.payload.amount #amount
-    ]
-    |> ExRLP.encode()
-  end
-  @spec rlp_encode(DataTx.t(OracleRegistrationTx.t())) :: binary()
-  def rlp_encode(%DataTx{type: OracleRegistrationTx} = tx) do
-    ttl_type = encode_ttl_type(tx.payload.ttl)
-
-    [
-      type_to_tag(OracleRegistrationTx),
-      get_version(OracleRegistrationTx),
-      tx.sender, #account
-      tx.nonce,  #nonce
-      # Subject to discuss/change:
-      # In Erlang core it is described as a UTF8 encoded String, but we have a map here
-      transform_item(tx.payload.query_format), #query_format/spec
-      # Subject to discuss/change:
-      # In Erlang core it is described as a UTF8 encoded String, but we have a map here
-      transform_item(tx.payload.response_format), #query_response/spec
-      tx.payload.query_fee, #query_fee
-      ttl_type, # ttl_type
-      tx.payload.ttl.ttl, #ttl_value
-      tx.fee #fee
-    ]
-    |> ExRLP.encode()
-  end
-  @spec rlp_encode(DataTx.t(OracleQueryTx.t())) :: binary()
-  def rlp_encode(%DataTx{type: OracleQueryTx} = tx) do
-    ttl_type_q = encode_ttl_type(tx.payload.query_ttl)
-    ttl_type_r = encode_ttl_type(tx.payload.response_ttl)
-
-    [
-      type_to_tag(OracleQueryTx),
-      get_version(OracleQueryTx),
-      tx.sender, #sender
-      tx.nonce,  #nonce
-      tx.payload.oracle_address, #oracle
-      # Subject to discuss/change:
-      # In Erlang core query_data is described as a binary,
-      # but not encoded "natively"(query_data in our case is a map)
-      transform_item(tx.payload.query_data), #query
-      tx.payload.query_fee, #query_fee
-      ttl_type_q, #query_ttl_type
-      tx.payload.query_ttl.ttl, #query_ttl_value
-      ttl_type_r, #response_ttl_type
-      tx.payload.response_ttl.ttl, #response_ttl_value
-      tx.fee #fee
-    ]
-    |> ExRLP.encode()
-  end
-  @spec rlp_encode(DataTx.t(OracleResponseTx.t())) :: binary()
-  def rlp_encode(%DataTx{type: OracleResponseTx} = tx) do
-
-    [
-      type_to_tag(OracleResponseTx),
-      get_version(OracleResponseTx),
-      tx.sender, #oracle? not confirmed
-      tx.nonce, #nonce
-      tx.payload.query_id, #query_id
-      transform_item(tx.payload.response), #response
-      tx.fee #fee
-    ]
-    |> ExRLP.encode()
-  end
-  @spec rlp_encode(Block.t()) :: binary()
-   def rlp_encode(%Block{} = block) do
-     header_bin = header_to_binary(block.header)
-     txs =
-     for tx <- block.txs do
-       rlp_encode(tx)
-     end
-     [
-       type_to_tag(Block),
-       block.header.version,
-       header_bin,
-       txs
-     ]
-     |>
-     ExRLP.encode
-   end
-  @spec rlp_encode(DataTx.t(OracleExtendTx.t())) :: binary()
-  def rlp_encode(%DataTx{type: OracleExtendTx} = tx) do
-    ttl_type = encode_ttl_type(tx.payload.ttl)
-
-    [
-      type_to_tag(OracleExtendTx),
-      get_version(OracleExtendTx),
-      tx.sender, #oracle? not confirmed
-      tx.nonce,  #nonce
-      ttl_type,  #ttl_type
-      tx.payload.ttl.ttl, #ttl_value
-      tx.fee #fee
-    ]
-    |> ExRLP.encode()
-  end
-
-  @spec rlp_encode(Account.t(), Wallet.pubkey()) :: binary()
-  def rlp_encode(%Account{} = account, pkey) do
-    [
-      type_to_tag(Account),
-      get_version(Account),
-      pkey, #pubkey ,
-      account.nonce, #nonce
-      account.last_updated,    #height
-      account.balance #balance
-    ]
-    |>
-    ExRLP.encode
-  end
-  @spec rlp_encode(Chainstate.t(), Wallet.pubkey(), atom()) :: binary()
-  def rlp_encode(%Chainstate{accounts: accounts}, pkey, :account_state) do
-    account_info = Account.get_account_state(accounts, pkey)
-
-    [
-      type_to_tag(Account),
-      get_version(Account),
-      pkey, #pubkey
-      account_info.nonce, #should be height but atm its nonce
-      account_info.balance #balance
-    ]
-    |> ExRLP.encode()
-  end
-  def rlp_encode(
-        %Chainstate{oracles: %{registered_oracles: registered_oracles}},
-        orc_owner,
-        :oracle_state
-      ) do
-    # list_of_formatted_data =
-    # Subject to discuss field "height_included" in our structures is not being encoded , erlang" core doesnt have this field
-    case Map.get(registered_oracles, orc_owner) do
-      nil ->
-        [
-          type_to_tag(Oracle),
-          get_version(Oracle),
-          orc_owner, # owner
-          transform_item(%{}), #query_format, should be a string(but we have a map)
-          transform_item(%{}), #response_format, should be a string (but we have a map)
-          0, #query_fee
-          0, #expires
-        ]
-
-      data ->
-        [
-          type_to_tag(Oracle),
-          get_version(Oracle),
-          orc_owner,  # owner
-          # Subject to discuss/change - In Erlang implementation its a UTF8 encoded string but in our case - map
-          transform_item(data.tx.query_format), #query_format, should be a string(but we have a map)
-          transform_item(data.tx.response_format), #response_format, should be a string (but we have a map)
-          data.tx.query_fee, #query_fee
-          # Subject to discuss/change - Erlang core has an integer field called expires, our case - map with ttl and its type :absolute/relative
-         # encode_ttl_type(data.tx.ttl),
-          data.tx.ttl.ttl  #"expires" doesnt exist at this moment,
-        ]
-    end
-    |> ExRLP.encode()
-  end
-  def rlp_encode(
-        %Chainstate{oracles: %{interaction_objects: interaction_objects}},
-        sender_address,
-        :oracle_query
-      ) do
-    case Map.get(interaction_objects, sender_address) do
-      nil ->
-        [
-          type_to_tag(OracleQuery), #type
-          get_version(OracleQuery), #ver
-          sender_address, #pubkey
-          0, #sender_nonce
-          <<0>>, # oracle_address
-          transform_item(%OracleQueryTx{}), #query
-          <<0>>, #has_response
-          transform_item(%OracleResponseTx{}), #response
-          0, #expires
-          0, #response_ttl
-          0, #fee
-        ]
-
-      data ->
-        [
-          type_to_tag(OracleQuery), #type
-          get_version(OracleQuery), #ver
-          sender_address, #pubkey
-          # Subject : We dont have sender_nonce field here,
-          data.query.oracle_address, # oracle_address
-          # Subject : Query: Equivalent here would be fully composed DataTx(OracleQueryTx) , not just OracleQueryTx
-          rlp_encode(transform_item(data.query)), #query
-          <<0>>,  # Subject: has_response field doesnt exist, will be hardcoded as <<0>>/false for now
-
-          # Subject: Response: Equivalent here would be fully composed DataTx(ResponseTx), not a map
-          rlp_encode(transform_item(data.query.response)), #response
-          data.query.query_ttl.ttl, # Subject: no field called "expires" here, should be calculated over ttl_type and ttl_value itself
-          data.query.response_ttl.ttl,
-          data.query.query_fee
-        ]
-    end
-    |> ExRLP.encode()
-  end
-  @spec header_to_binary((Header.t())) :: binary
-  def header_to_binary(%Header{} = header) do
-    pow_to_binary =
-    pow_to_binary(header.pow_evidence)
-     <<
-      header.version::64,
-      header.height::64,
-      header.prev_hash::binary-size(32),
-      header.txs_hash::binary-size(32),
-      header.root_hash::binary-size(32),
-      header.target::64,
-      pow_to_binary::binary-size(168),
-      header.nonce::64,
-      header.time::64
-    >>
-   end
-  def header_to_binary(_) do
-     :illegal_structure_serialization
-  end
-  @spec binary_to_header(binary()) :: Header.t()
-  def binary_to_header(binary) when is_binary(binary) do
-     <<version::64,
-       height::64,
-       prev_hash::binary-size(32),
-       txs_hash::binary-size(32),
-       root_hash::binary-size(32),
-       target::64,
-       pow_evidence_bin::binary-size(168),
-       nonce::64,
-       time::64
-     >> = binary
-    pow_evidence = binary_to_pow(pow_evidence_bin)
-    %Header
-     {
-      height: height,
-      nonce: nonce,
-      pow_evidence: pow_evidence,
-      prev_hash: prev_hash,
-      root_hash: root_hash,
-      target: target,
-      time: time,
-      txs_hash: txs_hash,
-      version: version
-     }
-  end
-  def binary_to_header(_) do
-     :illegal_Header_serialization
-  end
-
-  defp type_to_tag(Account), do: 10
-  defp type_to_tag(SignedTx), do: 11
-  defp type_to_tag(SpendTx), do: 12
-  defp type_to_tag(CoinbaseTx), do: 13
-  defp type_to_tag(Oracle), do: 20
-  defp type_to_tag(OracleQuery), do: 21
-  defp type_to_tag(OracleRegistrationTx), do: 22
-  defp type_to_tag(OracleQueryTx), do: 23
-  defp type_to_tag(OracleResponseTx), do: 24
-  defp type_to_tag(OracleExtendTx), do: 25
-  defp type_to_tag(Block), do: 100
-
-  defp tag_to_type(10), do: Account
-  defp tag_to_type(11), do: SignedTx
-  defp tag_to_type(12), do: SpendTx
-  defp tag_to_type(13), do: CoinbaseTx
-  defp tag_to_type(20), do: Oracle
-  defp tag_to_type(21), do: OracleQuery
-  defp tag_to_type(22), do: OracleRegistrationTx
-  defp tag_to_type(23), do: OracleQueryTx
-  defp tag_to_type(24), do: OracleResponseTx
-  defp tag_to_type(25), do: OracleExtendTx
-  defp tag_to_type(100), do: Block
-  @spec rlp_decode(binary()) :: {:ok, Account.t()} | Block.t() | DataTx.t()
-  def rlp_decode(values) when is_binary(values) do
-    [tag_bin, ver_bin | rest_data] = ExRLP.decode(values)
-    tag = transform_item(tag_bin, :int)
-    ver = transform_item(ver_bin, :int)
-
-    case tag_to_type(tag) do
-      Account ->
-        [pkey, nonce, height, balance] = rest_data
-        [pkey, transform_item(nonce, :int), transform_item(height, :int), transform_item(balance, :int)]
-        {:ok, %Account{balance: transform_item(balance, :int), last_updated: transform_item(height, :int), nonce: transform_item(nonce, :int)}}
-
-      SignedTx ->
-        [signatures, tx_data] = rest_data
-        decoded_signatures = for sig <- signatures do
-            ExRLP.decode(sig)
-          end
-         %SignedTx{data: rlp_decode(tx_data) , signature: decoded_signatures}
-
-      SpendTx ->
-        [sender, receiver, amount, fee, nonce] = rest_data
-
-        [
-          sender,
-          receiver,
-          transform_item(amount, :int),
-          transform_item(fee, :int),
-          transform_item(nonce, :int)
-        ]
-         DataTx.init(SpendTx,%{receiver: receiver, amount: transform_item(amount, :int), version: ver},sender,transform_item(fee, :int),transform_item(nonce, :int))
-
-      CoinbaseTx ->
-        [receiver, nonce, amount] = rest_data
-        [receiver, transform_item(nonce, :int), transform_item(amount, :int)]
-        %DataTx{fee: 0, nonce: transform_item(nonce, :int), payload: %SpendTx{amount: transform_item(amount, :int), receiver: receiver, version: ver}, sender: nil, type: SpendTx}
-        DataTx.init(SpendTx, %{
-          receiver: receiver,
-          amount: transform_item(amount, :int),
-          version: ver
-        }, 
-        nil,
-        0, 
-        transform_item(nonce, :int))
-      Oracle ->
-        [orc_owner, query_format, response_format, query_fee, ttl_type, ttl] = rest_data
-
-        [
-          orc_owner,
-          transform_item(query_format, :binary),
-          transform_item(response_format, :binary),
-          transform_item(query_fee, :int),
-          transform_item(ttl_type, :int),
-          transform_item(ttl, :int)
-        ]
-       # DataTx.init(Oracle,)
-
-      OracleQueryTx ->
-        [
-          sender,
-          nonce,
-          oracle_address,
-          query_data,
-          query_fee,
-          query_ttl_type,
-          query_ttl_value,
-          response_ttl_type,
-          response_ttl_value,
-          fee
-        ] = rest_data
-
-         [
-          sender,
-          transform_item(nonce, :int),
-          oracle_address,
-          query_data,
-          transform_item(query_fee, :int),
-          query_ttl_type,
-          transform_item(query_ttl_value, :int),
-          response_ttl_type,
-          transform_item(response_ttl_value, :int),
-          transform_item(fee, :int)
-        ]
-
-      OracleRegistrationTx ->
-        [sender, nonce, query_format, response_format, query_fee, ttl_type, ttl_value, fee] =
-          rest_data
-
-        [
-          sender,
-          transform_item(nonce, :int),
-          transform_item(query_format, :binary),
-          transform_item(response_format, :binary),
-          transform_item(query_fee, :int),
-          transform_item(ttl_type, :binary),
-          transform_item(ttl_value, :int),
-          transform_item(fee, :int)
-        ]
-
-      OracleResponseTx ->
-        [sender, nonce, query_id, response, fee] = rest_data
-
-        [
-          sender,
-          transform_item(nonce, :int),
-          transform_item(query_id, :binary),
-          transform_item(response, :binary),
-          transform_item(fee, :int)
-        ]
-
-      OracleExtendTx ->
-        [sender, nonce, ttl_type, ttl_value, fee] = rest_data
-
-         [
-          sender,
-          transform_item(nonce, :int),
-          transform_item(ttl_type, :binary),
-          transform_item(ttl_value, :int),
-          transform_item(fee, :int)
-         ]
-      Block ->
-        [header_bin, txs] = rest_data
-        txs_list =
-          for tx <- txs do
-            rlp_decode(tx)
-          end
-
-        Block.new(%{header: binary_to_header(header_bin),txs: txs_list})
-
-      _ ->
-        {:error, :illegal_serialization}
-    end
-  end
-  def rlp_decode(binary) when is_binary(binary) do
-    ExRLP.decode binary
-  end
-  def rlp_decode(value) do
-    value
-  end
- 
-  #  def rlp_decode(binary,pattern) do
-  #  end
-  defp transform_item(item) do
+  def transform_item(item) do
     :erlang.term_to_binary(item)
   end
 
-  defp transform_item(item, type) do
+  def transform_item(item, type) do
     case type do
       :int -> :binary.decode_unsigned(item)
       :binary -> :erlang.binary_to_term(item)
@@ -800,65 +344,10 @@ defmodule Aeutil.Serialization do
     end
   end
 
-  defp encode_ttl_type(%{type: type}) do
+  def encode_ttl_type(%{type: type}) do
     case type do
       :absolute -> 0
       :relative -> 1
     end
-  end
-  @spec pow_to_binary(list()) :: binary()
-  def pow_to_binary(pow) when is_list(pow) do
-    if Enum.count(pow) == 42 do
-      list_of_pows =
-        for evidence <- pow do
-          <<evidence::32>>
-        end
-        serialize_pow(:binary.list_to_bin(list_of_pows),<<>>)
-    else
-      List.duplicate(0, 42)
-    end
-  end
-  defp serialize_pow(pow, acc) when pow != <<>> do
-    <<elem ::binary-size(4), rest::binary>> = pow
-    acc = acc <> elem
-    serialize_pow(rest, acc)
-  end
-  defp serialize_pow(<<>>, acc) do
-    acc
-  end
-  @spec binary_to_pow(binary()) :: list() | {:error, atom()}
-  def binary_to_pow(<<pow_bin_list::binary-size(168)>>) do
-    deserialize_pow(pow_bin_list,[])
-  end
-  def binary_to_pow(_) do
-    {:error, :illegal_PoW_serialization}
-  end
-  defp deserialize_pow(<<pow::32,rest::binary>>, acc)  do
-    acc = acc ++ [pow]
-    deserialize_pow(rest, acc)
-  end
-  defp deserialize_pow(<<>>, acc) do
-    Enum.filter(acc, fn(x) -> is_integer(x) and x >= 0 end)
-  end
-  @spec nils_to_binary(SignedTx.t()) :: Map.t()
-  def nils_to_binary(tx) when is_map(tx) do
-    tx = remove_struct(tx)
-
-    Enum.reduce(tx, %{}, fn {k, v}, acc ->
-      Map.put(
-        acc,
-        k,
-        # Subject to discuss/change:
-        # Added to avoid nils from SpendTx(Coin-based),
-        # because CoinbaseTx doesnt have separate structure at this moment
-        case v do
-          %{} = v ->
-            nils_to_binary(v)
-
-          _ ->
-            v
-        end
-      )
-    end)
   end
 end
